# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Dernier Metro API
  version: 1.0.0
  description: API pour gérer les stations de métro
servers:
  - url: http://localhost:5000
    description: Serveur local
paths:
...
  /next-metro:
    get:
      summary: Obtenir l'heure du prochain métro
      parameters:
        - in: query
          name: station
          schema:
            type: string
          required: true
          description: Nom de la station
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  station:
                    type: string
                  line:
                    type: string
                  headwayMin:
                    type: integer
                  nextArrival:
                    type: string
                    example: "23:45"
        '400':
          description: Paramètre manquant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur interne
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /last-metro:
    get:
      summary: Obtenir l'heure du dernier métro
      parameters:
        - in: query
          name: station
          schema:
            type: string
          required: true
          description: Nom de la station
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  station:
                    type: string
                  lastMetro:
                    type: string
                  line:
                    type: string
                  tz:
                    type: string
        '400':
          description: Paramètre manquant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Station inconnue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur interne
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string